Python_Script/requirements.txt
matplotlib == 3.1.3
numpy == 1.18.1
pyserial == 3.4
tabulate == 0.8.7
